db: start fresh with an SQL dump
At the moment, only a handful of the required tables are created via
Knex. The rest of the SQL tables can only be found via a full SQL dump
that developers send each other (~13Go), which is not viable.

Start fresh with a recent SQL *schema* dump that loads all the tables
via Knex. This means the project can be fully boostrapped started
locally.

In the future, modifying the database should be done with Knex
migrations that alter the tables provided in this initial SQL file,
which itself should never be touched again.

Co-authored-by: denis baudot <denis.baudot@beta.gouv.fr>